Replace old editing interface with Editor.js
