+++ components/editors/article/EditorTools.jsx
@@ -1,0 +1,17 @@
+import Paragraph from '@editorjs/paragraph'
+import Header from '@editorjs/header'
+import SimpleImage from '@editorjs/simple-image'
+import NestedList from '@editorjs/nested-list'
+import Quote from '@editorjs/quote'
+import Delimiter from '@editorjs/delimiter'
+import Embed from '@editorjs/embed'
+
+export const EDITOR_TOOLS = {
+   paragraph: Paragraph,
+   header: Header,
+   list: NestedList,
+   simpleImage: SimpleImage,
+   delimiter: Delimiter,
+   quote: Quote,
+   embed: Embed,
+}
