+++ package.json
@@ -3,17 +3,15 @@
    "version": "0.1.0",
    "private": true,
    "dependencies": {
-      "@bomdi/codebox": "^2.0.0",
-      "@editorjs/checklist": "^1.4.0",
       "@editorjs/delimiter": "^1.3.0",
       "@editorjs/editorjs": "^2.26.5",
       "@editorjs/embed": "^2.5.3",
       "@editorjs/header": "^2.7.0",
-      "@editorjs/inline-code": "^1.4.0",
-      "@editorjs/link": "^2.5.0",
-      "@editorjs/list": "^1.8.0",
+      "@editorjs/nested-list": "^1.3.0",
+      "@editorjs/paragraph": "^2.9.0",
       "@editorjs/quote": "^2.5.0",
       "@editorjs/simple-image": "^1.5.1",
+      "editorjs-blocks-react-renderer": "^1.3.0",
       "mongodb": "^5.1.0",
       "next": "^13.1.1",
       "react": "^18.2.0",
