+++ pages/journal/[log].jsx
@@ -8,8 +8,9 @@ import ContentEditable from 'react-contenteditable'
 import { getLogsUris, getLog } from 'utils/logs'
 import { withRouter } from 'next/router'
 import dynamic from 'next/dynamic'
+import Blocks from 'editorjs-blocks-react-renderer'
 
-const Editor = dynamic(() => import('components/Editor'), {
+const Editor = dynamic(() => import('components/editors/article/Editor'), {
    ssr: false,
 })
 
@@ -26,14 +27,19 @@ class Log extends Component {
          subtitle: this.props.log.subtitle,
          titleImage: this.props.log.titleImage,
          titleImageAlt: this.props.log.titleImageAlt,
-         articleContent: this.props.log.articleContent || 'Write here',
+         paragraphs: this.props.log.paragraphs,
+         articleContent: this.props.log.articleContent,
          // Component state
          showPreview: true,
       }
    }
 
    setArticleContent = (articleContent) => {
+      console.log(`Current state: ${JSON.stringify(this.state.articleContent)}`)
+
       this.setState({ articleContent: articleContent })
+
+      console.log(`Current state: ${JSON.stringify(this.state.articleContent)}`)
    }
 
    togglePreview = (event) => {
@@ -123,49 +129,16 @@ class Log extends Component {
             </header>
 
             <main className={`${logStyles.article}`}>
-               <Editor
-                  data={this.state.articleContent}
-                  onChange={this.setArticleContent}
-                  holder='editorjs-container'
-               />
-            </main>
-
-            {/* <main className={`${logStyles.article}`}>
-               {this.state.paragraphs
-                  ? this.state.paragraphs.map((_paragraph, paragraphId) => {
-                       return (
-                          <div key={paragraphId}>
-                             <ContentEditable
-                                tagName='p'
-                                className={`${logStyles.paragraph}`}
-                                onChange={(event) => this.handleParagraphChange(event, paragraphId)}
-                                html={this.state.paragraphs[paragraphId]}
-                                disabled={this.state.showPreview}
-                             />
-
-                             {!this.state.showPreview && (
-                                <input
-                                   type='button'
-                                   className={`${editLogStyles.button} ${editLogStyles.editParagraphButton}`}
-                                   value='Remove paragraph'
-                                   onClick={() => this.removeParagraph(paragraphId)}
-                                />
-                             )}
-                          </div>
-                       )
-                    })
-                  : null}
-
-               {!this.state.showPreview && (
-                  <input
-                     type='button'
-                     id={`${editLogStyles.addParagraphButton}`}
-                     className={`${editLogStyles.button}`}
-                     value='Add new paragrpah'
-                     onClick={() => this.addParagraph()}
+               {!this.state.showPreview ? (
+                  <Editor
+                     data={this.state.articleContent}
+                     onChange={this.setArticleContent}
+                     holder='editorjs-container'
                   />
+               ) : (
+                  <Blocks data={this.state.articleContent} />
                )}
-            </main> */}
+            </main>
 
             {!this.state.showPreview && (
                <form id={`${editLogStyles.editPageForm}`}>
@@ -273,6 +246,7 @@ class Log extends Component {
          titleImage: this.state.titleImage,
          titleImageAlt: this.state.titleImageAlt,
          paragraphs: this.state.paragraphs,
+         articleContent: this.state.articleContent,
       })
    }
